Changed threshold-based artifact rejection
The threshold-based artifact rejection function (ft_artifact_threshold) was not performing as expect and causing errors.  Updated this to use own implementation of threshold-based artifact detection.  Trials above threshold are removed.  If a set proportion of trials are above threshold for a particular channel in the data then the channel is removed instead and recovered in a later step.